Change rating from int to double

+++ Sources/ResponseStructs.swift
@@ -51,7 +51,8 @@ public struct GreetedUser: Identifiable, Codable, Equatable, Hashable {
     public var isBlocked: Bool
     public var manualNotificationFrequency: NotificationFrequency
     public var automaticNotificationFrequency: NotificationFrequency
-    public var rating: Int?
+    /// 0.0 - 1.0 represents 100%
+    public var rating: Double?
     public var greetIDs: [UUID] = []
     public var metOnDate: Date?
 
@@ -62,7 +63,7 @@ public struct GreetedUser: Identifiable, Codable, Equatable, Hashable {
         isBlocked: Bool = false,
         manualNotificationFrequency: NotificationFrequency = .unrestricted,
         automaticNotificationFrequency: NotificationFrequency = .unrestricted,
-        rating: Int? = nil,
+        rating: Double? = nil,
         greetIDs: [UUID] = [],
         metOnDate: Date? = nil
     ) {
